Add a ThemePicker
Theme handling is moved from the `App` component to
a separate `ThemePicker` component. This component
can be used to switch between different themes and also
stores the selected theme to local storage so it is persisted
across sessions of a user.

Of corse this also adds the best theme in the history of
themes: Dracula.